+++ banyand/liaison/grpc/registry_test.go
@@ -182,11 +182,11 @@ func setupForRegistry() func() {
 	metricSvc := observability.NewMetricService(metaSvc, pipeline, "standalone", nil)
 
 	nr := grpc.NewLocalNodeRegistry()
-	tcp := grpc.NewServer(context.TODO(), pipeline, pipeline, metaSvc, grpc.NodeRegistries{
+	tcp := grpc.NewServer(context.TODO(), pipeline, pipeline, nil, metaSvc, grpc.NodeRegistries{
 		MeasureNodeRegistry:  nr,
 		StreamNodeRegistry:   nr,
 		PropertyNodeRegistry: nr,
-	}, metricSvc)
+	}, metricSvc, nil)
 	preloadStreamSvc := &preloadStreamService{metaSvc: metaSvc}
 	var flags []string
 	metaPath, metaDeferFunc, err := test.NewSpace()